fix tests broken by moving testData
