+++ .codeclimate.yml
@@ -26,6 +26,6 @@ ratings:
   - "**.rb"
 exclude_paths:
 - "**/__tests__/**"
-- "src/common/__tests__/testData.js"
+- "src/common/testData.js"
 - "docs/**"
 JavaScript: true
